pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pit0500 <pietro.macri2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-cors-headers = "^4.3.1"
coverage = "^7.3.2"
coreapi = "^2.3.3"
djangorestframework = "^3.14.0"
pyyaml = "^6.0.1"
pytest-django = "^4.7.0"
mixer = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
